requirements.txt
flet
pydantic
pydantic[email]
pydantic-extra-types
phonenumbers
httpx
python-dotenv
